fix: Correctly process Facebook OAuth JSON response
For version 2.3 of the Graph API, the response when exchanging
the short-lived token for the long-lived token is now in JSON format.
This change takes that into account. It also explicitly defines the API
version, rather than leaving it to default to the oldest available. It was
this that previous left the error undetected.