fix: Use dictionary get() method instead of []
If there is no user logged in, the gdisconnect() will
raise a KeyError as the access_token is attempted
to be accessed directly.

Use of the get() method is shown in the video and
the original code, when the whole credentials object
was being stored.

Also removed some tab characters in gdisconnect()